requirements.txt
flask==3.*
mysql-connector-python==8.*
python-dotenv==1.*
openai==1.*
